Remove argparser.py & move parse_cli_args -> helpers
